\name{Converge-class}
\Rdversion{1.1}
\docType{class}
\alias{Converge-class}
\alias{alias}
\title{Class \code{"Converge"} }
\description{
Contain informations about the convergence and the number of iterations
of the algorithm.
}
\section{Slots}{
  \describe{
    \item{\code{nIter}:}{Number of iterations of the algorithm.}
    \item{\code{convergence}:}{ \code{"logical"} if [TRUE] then the algorithm met convergence. }
  }
}
\note{
  Meant to be used internally.
}
