\name{GlmCluster-methods}
\docType{methods}
\alias{plot,GlmCluster-method}
\title{Plot the main trajectories}
\description{
 Plot the main trajectories of each cluster and print the proportions of
 each one of it.
}
\section{Methods}{
\code{plot(GlmCluster)} : Plot the main trajectories.
}  
\seealso{
  \link{kmlCov}
    \link{glmClust}
  }
