\name{plot-methods}
\docType{methods}
\alias{plot,KmlCovList-method}
\title{Plot the main trajectories}
\description{
Plot the main trajectories of an object of class \link{KmlCovList} one
by one and ask the user to plot the next diagram.
}
\section{Methods}{
\code{plot(KmlCovList)} : Plot the main trajectories.
}