\name{majIndica}
\alias{majIndica}
\title{Calculate an indicator vector}
\usage{
  majIndica(aff_obs, itrClust)
}
\arguments{
  \item{aff_obs}{Vector of [integer].}

  \item{itrClust}{Number of current cluster.}
}
\value{
  An indicator vector of the belonging to a cluster.
}
\description{
  Calculate and return an indicator vector.
}
\note{
  Meant to be used internally.
}

